Move drop logic for relationships to diagram package
Since Relationship is a core construct now.
